requirements.txt
annotated-types==0.7.0
black==25.1.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
idna==3.10
markdown-it-py==4.0.0
mdurl==0.1.2
msgpack==1.0.2
mypy_extensions==1.1.0
numpy==2.3.2
packaging==25.0
pandas==2.3.2
pathspec==0.12.1
platformdirs==4.4.0
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-json-logger==3.3.0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.5
rich==14.1.0
setuptools==80.9.0
shellingham==1.5.4
signalrcore==0.9.5
six==1.17.0
typer==0.17.4
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
websocket-client==1.0.0
wheel==0.45.1
